deploy/requirements.txt
streamlit
requests
streamlit-authenticator
python-dotenv